feat : 펫 상세조회 추가, 삭제 시 validation 처리 추가, s3 파일 삭제 추가

+++ src/app/MyPet/petService.js
@@ -5,7 +5,7 @@ const petDao = require("./petDao");
 const baseResponse = require("../../../config/baseResponseStatus");
 const { response } = require("../../../config/response");
 const { errResponse } = require("../../../config/response");
-
+const s3 = require('../../../config/s3');
 const jwt = require("jsonwebtoken");
 const crypto = require("crypto");
 const { connect } = require("http2");
@@ -36,13 +36,30 @@ exports.deletePet = async function(userId, petId) {
         // if (userRows.length < 1) return errResponse(baseResponse.USER_ID_NOT_EXIST);
 
         // 펫 DB에 있는지 조회 
+        const petResult = await petProvider.retrievePetById(petId);
+        if (petResult.length < 1) return errResponse(baseResponse.PET_ID_NOT_EXIST);
+        else {
+            const petImage = petResult[0].profile_image.split('/')[4];
+            //console.log(petImage);
+            const connection = await pool.getConnection(async(conn) => conn);
 
-        const connection = await pool.getConnection(async(conn) => conn);
+            const deletePet = await petDao.deletePet(connection, petId);
 
-        const petResult = await petDao.deletePet(connection, petId);
-        connection.release();
+            s3.deleteObject({
+                Bucket: 'pindergarten/pet', // 사용자 버킷 이름
+                Key: petImage
+            }, function(err, data) {
+                if (err) {
+                    console.log('aws s3 delete error')
+                } else {
+                    console.log('aws s3 delete success');
+                }
+            });
+            connection.release();
 
-        return response(baseResponse.SUCCESS);
+            return response(baseResponse.SUCCESS);
+        }
+
     } catch (err) {
         logger.error(`APP - deletePet Service error\n: ${err.message}`);
         return errResponse(baseResponse.DB_ERROR);
